% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{activate,HCABrowser-method}
\alias{activate,HCABrowser-method}
\title{Activate files or bundles of Human}
\usage{
\S4method{activate}{HCABrowser}(hca, what = c("bundles", "files"))
}
\arguments{
\item{hca}{An HCABrowser object}

\item{what}{Either "bundles" or "files". Deterimines whether bundles or files
should be shown.}
}
\description{
Activate files or bundles of Human
}
