% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{activate-HCABrowser}
\alias{activate-HCABrowser}
\title{Activate files or bundles of Human}
\usage{
\S4method{activate}{HCABrowser}(hca, what = c("bundles", "files"))
}
\description{
Activate files or bundles of Human
}
