% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{activate-HumanCellAtlas}
\alias{activate-HumanCellAtlas}
\alias{activate.HumanCellAtlas}
\title{Activate files or bundles of Human}
\usage{
\method{activate}{HumanCellAtlas}(object, type = c("bundles", "files"),
  ...)
}
\description{
Activate files or bundles of Human
}
