% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{activate}
\alias{activate}
\alias{activate,HCABrowser-method}
\title{Activate files or bundles of HCABrowser Object}
\usage{
\S4method{activate}{HCABrowser}(hca, what = c("bundles", "files"))
}
\arguments{
\item{hca}{An HCABrowser object}

\item{what}{Either "bundles" or "files". Deterimines whether bundles or files
should be shown.}
}
\value{
An HCABrowser with the selected activation
}
\description{
Activate files or bundles of HCABrowser Object
}
