% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utilities.R
\docType{methods}
\name{availableFields,HumanCellAtlas-method}
\alias{availableFields,HumanCellAtlas-method}
\title{List all values for certain fields}
\usage{
\S4method{availableFields}{HumanCellAtlas}(hca, fields = c())
}
\description{
List all values for certain fields
}
