% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{deleteBundle,character-method}
\alias{deleteBundle,character-method}
\title{Delete a bundle or a specific bundle version}
\usage{
\S4method{deleteBundle}{character}(uuid, replica = c("aws", "gcp", "azure"),
  version = NULL, reason = NULL,
  url = "https://dss.data.humancellatlas.org/v1")
}
\description{
Delete a bundle or a specific bundle version
}
