% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{deleteCollection,character-method}
\alias{deleteCollection,character-method}
\title{Delete a collection}
\usage{
\S4method{deleteCollection}{character}(uuid, replica = c("aws", "gcp",
  "azure"), url = "https://dss.data.humancellatlas.org/v1")
}
\description{
Delete a collection
}
