% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{deleteSubscription,character-method}
\alias{deleteSubscription,character-method}
\title{Delete an event subscription}
\usage{
\S4method{deleteSubscription}{character}(uuid, replica = c("aws", "gcp",
  "azure"), url = "https://dss.integration.data.humancellatlas.org/v1")
}
\description{
Delete an event subscription
}
