% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\name{deprecated}
\alias{deprecated}
\alias{getBundle}
\title{Deprecated functions in the HCABrowser package}
\usage{
getBundle(x, uuid, version = NULL, replica = "aws")
}
\description{
\code{getBundle()} is deprecated, use \code{checkoutBundle()}
}
\keyword{internal}
