% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{downloadHCA,HCABrowser-method}
\alias{downloadHCA,HCABrowser-method}
\title{Download results from a HCABrowser query}
\usage{
\S4method{downloadHCA}{HCABrowser}(hca, n)
}
\arguments{
\item{hca}{A HCABrowser object}

\item{n}{integer(1) number of bundles to download}
}
\value{
a tibble all bundles obtained from download
}
\description{
Download results from a HCABrowser query
}
\examples{

hca <- HCABrowser()
res <- hca \%>\% downloadHCA(n = 24)
res

}
