% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{downloadHCA,HumanCellAtlas-method}
\alias{downloadHCA,HumanCellAtlas-method}
\title{Download results from a HumanCellAtlas query}
\usage{
\S4method{downloadHCA}{HumanCellAtlas}(hca, ..., n)
}
\arguments{
\item{hca}{A HumanCellAtlas object}

\item{n}{integer(1) number of bundles to download}
}
\value{
a tibble all bundles obtained from download
}
\description{
Download results from a HumanCellAtlas query
}
\examples{

hca <- HumanCellAtlas()
res <- hca \%>\% downloadHCA(n = 24)
res

}
