% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utilities.R
\name{filter.HumanCellAtlas}
\alias{filter.HumanCellAtlas}
\title{Filter HumanCellAtlas objects}
\usage{
\method{filter}{HumanCellAtlas}(hca, ...)
}
\arguments{
\item{hca}{a HumanCellAtlas object to perform a query on.}

\item{...}{further argument to be tranlated into a query to select from.
These arguments can be passed in two ways, either as a single expression or
as a series of expressions that are to be seperated by commas.}
}
\value{
a HumanCellAtlas object containing the resulting query.
}
\description{
Filter HumanCellAtlas objects
}
\examples{

hca <- HumanCellAtlas()
hca2 <- hca \%>\% filter()
hca2

hca3 <- hca \%>\% filter()
hca3


}
