% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{getBundle,character-method}
\alias{getBundle,character-method}
\title{Retrieve a bundle given a UUID and optionally a version}
\usage{
\S4method{getBundle}{character}(uuid, replica = c("aws", "gcp", "azure"),
  version = NULL, directurls = NULL, presignedurls = FALSE,
  token = NULL, url = "https://dss.integration.data.humancellatlas.org/v1")
}
\description{
Retrieve a bundle given a UUID and optionally a version
}
