% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{getBundlesCheckout,character-method}
\alias{getBundlesCheckout,character-method}
\title{Check the status of a checkout request}
\usage{
\S4method{getBundlesCheckout}{character}(checkout_job_id, replica = c("aws",
  "gcp", "azure"), url = "https://dss.data.humancellatlas.org/")
}
\description{
Check the status of a checkout request
}
