% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{getCollection,character-method}
\alias{getCollection,character-method}
\title{Retrieve a collection given a UUID}
\usage{
\S4method{getCollection}{character}(uuid, replica = c("aws", "gcp", "azure"),
  version = NULL, url = "https://dss.data.humancellatlas.org/v1")
}
\description{
Retrieve a collection given a UUID
}
