% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{getEsQuery,HCABrowser-method}
\alias{getEsQuery,HCABrowser-method}
\title{Get Elastic Search query as JSON}
\usage{
\S4method{getEsQuery}{HCABrowser}(object)
}
\arguments{
\item{hca}{An HCABrowser object}
}
\value{
A json object of the elastic search query in the HCABrowser object
}
\description{
Get Elastic Search query as JSON
}
