% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{getFile,character-method}
\alias{getFile,character-method}
\title{Retrieve a file given a UUID and optionally a version}
\usage{
\S4method{getFile}{character}(uuid, replica = c("aws", "gcp", "azure"),
  token = NULL, version = NULL,
  url = "https://dss.integration.data.humancellatlas.org/v1")
}
\description{
Retrieve a file given a UUID and optionally a version
}
