% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{getSubscription,character-method}
\alias{getSubscription,character-method}
\title{Retrieve an event subscription given a UUID}
\usage{
\S4method{getSubscription}{character}(uuid, replica = c("aws", "gcp",
  "azure"), url = "https://dss.data.humancellatlas.org/v1")
}
\description{
Retrieve an event subscription given a UUID
}
