% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{getSubscriptions,character-method}
\alias{getSubscriptions,character-method}
\title{Retrieve a user's event Subscription}
\usage{
\S4method{getSubscriptions}{character}(replica = c("aws", "gcp", "azure"),
  url = "https://dss.data.humancellatlas.org/v1")
}
\description{
Retrieve a user's event Subscription
}
