% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\name{hca-api-methods}
\alias{hca-api-methods}
\alias{getBundleCheckout}
\alias{getBundle}
\alias{checkoutBundle}
\alias{getFile}
\alias{headFile}
\alias{getBundleCheckout,HCABrowser-method}
\alias{getBundle,HCABrowser-method}
\alias{checkoutBundle,HCABrowser-method}
\alias{getFile,HCABrowser-method}
\alias{headFile,HCABrowser-method}
\title{HCA API methods}
\usage{
getBundleCheckout(x, ...)
getBundle(x, ...)
checkoutBundle(x, ...)
getFile(x, ...)
headFile(x, ...)

\S4method{getBundleCheckout}{HCABrowser}(x, replica = c("aws", "gcp"), checkout_job_id)

\S4method{getBundle}{HCABrowser}(
  x,
  uuid,
  version,
  replica = c("aws", "gcp"),
  directurls,
  presignedurls,
  token,
  per_page = 500
)

\S4method{checkoutBundle}{HCABrowser}(x, uuid, version, replica = c("aws", "gcp"), json_request_body)

\S4method{getFile}{HCABrowser}(x, uuid, replica = c("aws", "gcp"), version, token)

\S4method{headFile}{HCABrowser}(x, uuid, replica = c("aws", "gcp"), version)
}
\arguments{
\item{x}{An HCABrowser object that is the subject of the request.}

\item{replica}{character(1). A replica to fetch form. Can either be
set to "aws", "gcp", or "azure".  DEFAULT is "aws".}

\item{checkout_job_id}{character(1). A RFC4122-complliant ID for the checkout
job request.}

\item{uuid}{character(1). A RFC4122-compliant ID for the bundle.}

\item{version}{character(1). Timestamp of bundle creation in RFC3339.}

\item{directurls}{logical(1). Include direct-access URLs in the response.
This is mutually exclusive with the \code{presignedurls} parameter.  DEFAULT
is \code{NULL}.}

\item{presignedurls}{logical(1). Include presigned URLs in the response. This
is mutually exclusive with the directurls parameter.}

\item{token}{\code{Token}. Token to manage retries. End users constructing
queries should not set this parameter. Use \code{get_token()} to generate.}

\item{per_page}{numeric(1). Max number of results to return per page.}

\item{json_request_body}{character(1) of a json query to be executed.}

\item{...}{Additional arguments.}
}
\value{
an HCABrowser object
}
\description{
Methods to access the Human Cell Atlas's Data Coordination Platform (HCA DCP)
by means of the platform's REST API.
}
\examples{
hca <- HCABrowser()
#addmore


}
\author{
Daniel Van Twisk
}
