% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HCABrowser-class}
\alias{HCABrowser-class}
\alias{.HCABrowser}
\title{The HCABrowser Class}
\arguments{
\item{per_page}{numeric(1) numbers of pages to view at a time.}

\item{host}{character(1) path to hca-dcp server}

\item{api_url}{character(1) path to schema}
}
\value{
An HCABrowser object.
}
\description{
The HCABrowser Class
}
\examples{
 hca <- HCABrowser()
 hca

}
\author{
Daniel Van Twisk
}
