% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{HCABrowser}
\alias{HCABrowser}
\title{The HCABrowser Class}
\usage{
HCABrowser(per_page = 10)
}
\arguments{
\item{per_page}{numeric(1) numbers of pages to view at a time.}

\item{host}{character(1) path to hca-dcp server}

\item{api_url}{character(1) path to schema}
}
\value{
An HCABrowser object.
}
\description{
The HCABrowser Class
}
\examples{
 hca <- HCABrowser()
 hca

}
\author{
Daniel Van Twisk
}
