% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{headFile,character-method}
\alias{headFile,character-method}
\title{Retrieve a file's metadata given an UUID and optionally a version}
\usage{
\S4method{headFile}{character}(uuid, replica = c("aws", "gcp", "azure"),
  version = NULL, url = "https://dss.data.humancellatlas.org/v1")
}
\description{
Retrieve a file's metadata given an UUID and optionally a version
}
