% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azul_api.R, R/hca-api.R
\docType{methods}
\name{nextResults}
\alias{nextResults}
\alias{nextResults,ProjectBrowser-method}
\alias{nextResults,HCABrowser-method}
\title{Next Results}
\usage{
\S4method{nextResults}{ProjectBrowser}(result)

\S4method{nextResults}{HCABrowser}(result)
}
\arguments{
\item{result}{A HCABrowser object that has further bundles to display.}
}
\value{
A ProjectBrowser object that displays the next results

A Human Cell Atlas object that displays the next results
}
\description{
Fetch the next set of bundles from a ProjectBrowser Object

Fetch the next set of bundles from a Human Cell Atlas Object
}
\examples{

hca <- ProjectBrowser()
hca <- nextResults(hca)
hca


hca <- HCABrowser()
hca <- nextResults(hca)
hca

}
\author{
Daniel Van Twisk

Daniel Van Twisk
}
