% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\name{parseToSearchResults}
\alias{parseToSearchResults}
\title{Parse results from a search query to a Search Results object}
\usage{
parseToSearchResults(res)
}
\arguments{
\item{res}{the results from an HCA query.}
}
\value{
a SearchResults object.
}
\description{
Parse results from a search query to a Search Results object
}
