% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{patchCollection,character-method}
\alias{patchCollection,character-method}
\title{Update a collection}
\usage{
\S4method{patchCollection}{character}(uuid, replica = c("aws", "gcp",
  "azure"), version, add_contents, description, details, name, remove_contents,
  url = "https://dss.data.humancellatlas.org/v1")
}
\description{
Update a collection
}
