% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{per_page,HumanCellAtlas-method}
\alias{per_page,HumanCellAtlas-method}
\title{Set per_page argument of HumanCellAtlas object}
\usage{
\S4method{per_page}{HumanCellAtlas}(hca, n)
}
\arguments{
\item{hca}{a HumanCellAtlas object}

\item{n}{the new per_page value}
}
\value{
a HumanCellAtlas with updated per_page value
}
\description{
note that no more than 10 pages can be displayed at once
}
\examples{

hca <- HumanCellAtlas()
hca <- per_page(hca, 5)
hca

}
