% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{per_page}
\alias{per_page}
\alias{per_page,HCABrowser-method}
\title{Set per_page argument of an HCABrowser object}
\usage{
\S4method{per_page}{HCABrowser}(x, n)
}
\arguments{
\item{x}{an HCABrowser object}

\item{n}{the new per_page value}
}
\value{
a HCABrowser with updated per_page value
}
\description{
note that no more than 10 pages can be displayed at once
}
\examples{
x <- HCABrowser()
#x <- per_page(x, 5)
x
}
