% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{postBundlesCheckout,character-method}
\alias{postBundlesCheckout,character-method}
\title{Check out a bundle to DSS-namaged or user-managed cloud object storage
destination}
\usage{
\S4method{postBundlesCheckout}{character}(uuid, replica = c("aws", "gcp",
  "azure"), destination = NULL, email = NULL,
  url = "https://dss.data.humancellatlas.org/v1")
}
\description{
Check out a bundle to DSS-namaged or user-managed cloud object storage
destination
}
