% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ProjectBrowser-class}
\alias{ProjectBrowser-class}
\alias{.ProjectBrowser}
\title{The Project Browser Class}
\arguments{
\item{url}{character(1) The url of the Human Cell Atlas.}
}
\description{
A still tentative class that displays Human Cell Atlas
 information by projects.
}
\author{
Daniel Van Twisk
}
