% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{pullBundles,HumanCellAtlas-method}
\alias{pullBundles,HumanCellAtlas-method}
\title{Obtain bunlde fqids from a HumanCellAtlas object}
\usage{
\S4method{pullBundles}{HumanCellAtlas}(hca, ...)
}
\arguments{
\item{hca}{A HumanCellAtlas object}
}
\value{
character(1) of bundle fqids
}
\description{
Obtain bunlde fqids from a HumanCellAtlas object
}
\examples{

hca <- HumanCellAtlas()
hca <- hca \%>\% pullBundles

}
