% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{pullBundles}
\alias{pullBundles}
\alias{pullBundles,HCABrowser-method}
\title{Obtain bundle fqids from a HCABrowser object}
\usage{
\S4method{pullBundles}{HCABrowser}(hca, n = hca@per_page)
}
\arguments{
\item{hca}{A HCABrowser object}

\item{n}{integer(1) number of bundle fqids to pull}
}
\value{
character(1) of bundle fqids
}
\description{
Obtain bundle fqids from a HCABrowser object
}
\examples{
hca <- HCABrowser()
hca <- hca \%>\% pullBundles
}
