% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{putBundle,character-method}
\alias{putBundle,character-method}
\title{Create a bundle}
\usage{
\S4method{putBundle}{character}(uuid, replica = c("aws", "gcp", "azure"),
  version = NULL, creator_uid, files,
  url = "https://dss.data.humancellatlas.org/v1")
}
\description{
Create a bundle
}
