% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{putCollection,character-method}
\alias{putCollection,character-method}
\title{Create a collection}
\usage{
\S4method{putCollection}{character}(uuid, replica = c("aws", "gcp", "azure"),
  version, contents, description, details, name,
  url = "https://dss.integration.data.humancellatlas.org/v1")
}
\description{
Create a collection
}
