% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{putFile,character-method}
\alias{putFile,character-method}
\title{Create a new version of a file}
\usage{
\S4method{putFile}{character}(uuid, creator_uid, source_url, version = NULL,
  url = "https://dss.integration.data.humancellatlas.org/v1")
}
\description{
Create a new version of a file
}
