% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{putSubscription,character-method}
\alias{putSubscription,character-method}
\title{Creates an event subscription}
\usage{
\S4method{putSubscription}{character}(replica = c("aws", "gcp", "azure"),
  attachments, callback_url, encoding, es_query, form_fields, hmac_key_id,
  hmac_secret_key, method, payload_form_field,
  url = "https://dss.integration.data.humancellatlas.org/v1")
}
\description{
Creates an event subscription
}
