% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{resetEsQuery,HCABrowser-method}
\alias{resetEsQuery,HCABrowser-method}
\title{Reset the query of a HCABrowser object to the default query}
\usage{
\S4method{resetEsQuery}{HCABrowser}(hca)
}
\arguments{
\item{hca}{A HCABrowser object}
}
\value{
A HCABrowser object with the search reset
}
\description{
Reset the query of a HCABrowser object to the default query
}
\examples{
\dontrun{
hca <- HCABrowser()
hca <- hca \%>\% filter(organ.text == brain)
hca <- hca \%>\% filter(organ.text != brain)
hca <- hca \%>\% resetEsQuery
hca
}
}
