% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{resetEsQuery,HumanCellAtlas-method}
\alias{resetEsQuery,HumanCellAtlas-method}
\title{Reset the query of a HumanCellAtlas object to the default query}
\usage{
\S4method{resetEsQuery}{HumanCellAtlas}(hca, ...)
}
\arguments{
\item{hca}{A HumanCellAtlas object}
}
\value{
A HumanCellAtlas object with the search reset
}
\description{
Reset the query of a HumanCellAtlas object to the default query
}
\examples{

hca <- HumanCellAtlas()
hca <- hca \%>\% filter(organ.text == brain)
hca <- hca \%>\% filter(organ.text != brain)
hca <- hca \%>\% resetEsQuery
hca

}
