% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{resetEsQuery}
\alias{resetEsQuery}
\alias{resetEsQuery,HCABrowser-method}
\title{Reset the query of a HCABrowser object to the default query}
\usage{
\S4method{resetEsQuery}{HCABrowser}(x)
}
\arguments{
\item{x}{A HCABrowser object}
}
\value{
A HCABrowser object with the search reset
}
\description{
Reset the query of a HCABrowser object to the default query
}
\examples{
x <- HCABrowser()
x <- x \%>\% filter(organ.text == brain)
x <- x \%>\% filter(organ.text != brain)
x <- x \%>\% resetEsQuery
x
}
