% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{results,HumanCellAtlas-method}
\alias{results,HumanCellAtlas-method}
\title{Obtain search results from a Human Cell Atlas Object}
\usage{
\S4method{results}{HumanCellAtlas}(object)
}
\arguments{
\item{object}{A Human Cell Atlas object}
}
\value{
a tibble
}
\description{
Returns a tibble either showing bundles or files based on whichever is
 activated.
}
