% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{results,SearchResult-method}
\alias{results,SearchResult-method}
\title{Get results of SearchResult object}
\usage{
\S4method{results}{SearchResult}(object)
}
\arguments{
\item{object}{A Searchresult to obtain the result slot value from}
}
\value{
tibble of the results of the HCABrowser query
}
\description{
Get results of SearchResult object
}
