% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{results}
\alias{results}
\alias{results,SearchResult-method}
\title{Get results of SearchResult object}
\usage{
\S4method{results}{SearchResult}(x)
}
\arguments{
\item{x}{A Searchresult to obtain the result slot value from}
}
\value{
list of the results of the HCABrowser query
}
\description{
Get results of SearchResult object
}
\examples{
 sr <- new("SearchResult")
 results(sr)

}
