% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchResult.R
\docType{class}
\name{SearchResult-class}
\alias{SearchResult-class}
\alias{.SearchResult}
\alias{nextResults}
\alias{results}
\alias{as_tibble.SearchResult}
\alias{show,SearchResult-method}
\title{Results from searching the HCA}
\usage{
nextResults(x)

results(x)

\method{as_tibble}{SearchResult}(x, ...)

\S4method{show}{SearchResult}(object)
}
\arguments{
\item{x}{A SearchResult returned by `searchBundles()`.}

\item{...}{additional parameters (unused)}

\item{object}{a SearchResult object.}
}
\value{
`nextResults()` returns a `SearchResult` representing the
    next 'page' of results from the original query.

`results()` returns a list-of-lists, with each element
    representing a distinct results of the HCABrowser query.

The `as_tibble` method for `SearchResult` returns a tibble
    representing the search result, with each row corresponding to
    a distinct result.
}
\description{
A 'SearchResult' is returned by `searchBundles()`, and
    represents the results for the search. Use `results()` to view
    current results, and `nextResults()` to page through results.
}
\examples{
bundles <- searchBundles(HCABrowser())  # Find all bundles...
bundles                                 # ... but retrieve first 100

nextResults(bundles)                    # ... and then next 100

bundles \%>\%
    ## results() returns a list-of-lists
    results() \%>\%
    ## this is the structure of the first result
    head(1) \%>\%
    str()

bundles \%>\% as_tibble()

}
