% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utilities.R
\name{select.HCABrowser}
\alias{select.HCABrowser}
\title{Select fields from a HCABrowser object}
\usage{
\method{select}{HCABrowser}(.data, ..., .output_format = c("raw",
  "summary"))
}
\arguments{
\item{.data}{an HCABrowser object to perform a selection on}

\item{...}{further argument to be tranlated into an expression to select from.
These arguments can be passed in two ways, either as a character vector or
as a series of expressions that are the fields that are to be selected
seperated by commas.}

\item{.output_format}{unused.}
}
\value{
a HCABrowser object containing the results of the selection.
}
\description{
Select fields from a HCABrowser object
}
\examples{
hca <- HCABrowser()
hca2 <- hca \%>\% select('paired_end')
hca2

hca3 <- hca \%>\% select(c('organ.text', 'paired_end'))
hca3
}
