% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utilities.R
\name{select.HumanCellAtlas}
\alias{select.HumanCellAtlas}
\title{Select fields from a HumanCellAtlas object}
\usage{
\method{select}{HumanCellAtlas}(hca, ..., .search = TRUE)
}
\arguments{
\item{hca}{a HumanCellAtlas object to perform a selection on}

\item{...}{further argument to be tranlated into an expression to select from.
These arguments can be passed in two ways, either as a character vector or
as a series of expressions that are the fields that are to be selected
seperated by commas.}
}
\value{
a HumanCellAtlas object containing the results of the selection.
}
\description{
Select fields from a HumanCellAtlas object
}
\examples{

hca <- HumanCellAtlas()
hca2 <- hca \%>\% select(paired_end)
hca2

hca3 <- hca \%>\% select(c('organ.text', 'paired_end'))
hca3

}
