% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,HCABrowser-method}
\alias{show,HCABrowser-method}
\title{Show HCABrowser object}
\usage{
\S4method{show}{HCABrowser}(object)
}
\arguments{
\item{object}{An HCAbrowser object to show}
}
\value{
outputs a text represntation of the object
}
\description{
Show HCABrowser object
}
