% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{show,ProjectBrowser-method}
\alias{show,ProjectBrowser-method}
\title{Show ProjectBrowser}
\usage{
\S4method{show}{ProjectBrowser}(object)
}
\arguments{
\item{object}{a ProjectBrowser object to show}
}
\value{
outputs a text represntation of the object
}
\description{
Show ProjectBrowser
}
