% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,SearchResult-method}
\alias{show,SearchResult-method}
\title{Show Search Result}
\usage{
\S4method{show}{SearchResult}(object)
}
\arguments{
\item{object}{a SearchResult object to show}
}
\value{
outputs a text represntation of the object
}
\description{
Show Search Result
}
