% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{showBundles,HCABrowser-method}
\alias{showBundles,HCABrowser-method}
\title{Obtain all bundles from an hca object using there bundle fqids}
\usage{
\S4method{showBundles}{HCABrowser}(hca, bundle_fqids)
}
\arguments{
\item{hca}{a HCABrowser object to search for bundles on.}

\item{bundle_fqids}{a character()}
}
\value{
A HCABrowser object displaying the selected bundles
}
\description{
Obtain all bundles from an hca object using there bundle fqids
}
\examples{
\dontrun{
hca <- HCABrowser()
hca_bundles <- hca \%>\% pullBundles('')
hca_bundles
}
}
