% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{showBundles,HumanCellAtlas-method}
\alias{showBundles,HumanCellAtlas-method}
\title{Obtain all bundles from an hca object using there bundle fqids}
\usage{
\S4method{showBundles}{HumanCellAtlas}(hca, bundle_fqids, ...)
}
\arguments{
\item{hca}{a HumanCellAtlas object to search for bundles on.}

\item{bundle_fqids}{a character()}
}
\value{
A HumanCellAtlas object displaying the selected bundles
}
\description{
Obtain all bundles from an hca object using there bundle fqids
}
\examples{

hca <- HumanCellAtlas()
hca_bundles <- hca \%>\% pullBundles('')
hca_bundles

}
