% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utilities.R
\docType{methods}
\name{supportedFields,HumanCellAtlas-method}
\alias{supportedFields,HumanCellAtlas-method}
\title{List supported fields of an HumanCellAtlas object}
\usage{
\S4method{supportedFields}{HumanCellAtlas}(hca)
}
\description{
List supported fields of an HumanCellAtlas object
}
