% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{undoEsQuery,HumanCellAtlas-method}
\alias{undoEsQuery,HumanCellAtlas-method}
\title{Undo previous filter queries on a HumanCellAtlas object}
\usage{
\S4method{undoEsQuery}{HumanCellAtlas}(hca, ..., n = 1L)
}
\arguments{
\item{hca}{A HumanCellAtlas object}

\item{n}{integer(1) the number of filter queries to undo}
}
\value{
A HumanCellAtlas object with n fewer queries
}
\description{
Undo previous filter queries on a HumanCellAtlas object
}
\examples{

hca <- HumanCellAtlas()
hca <- hca \%>\% filter(organ.text == brain)
hca <- hca \%>\% filter(organ.text == heart)
hca <- hca \%>\% filter(organ.text != brain)
hca <- hca \%>\% undoEsquery(n = 2)
hca

}
