% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{undoEsQuery}
\alias{undoEsQuery}
\alias{undoEsQuery,HCABrowser-method}
\title{Undo previous filter queries on a HCABrowser object}
\usage{
\S4method{undoEsQuery}{HCABrowser}(x, n = 1L)
}
\arguments{
\item{x}{A HCABrowser object}

\item{n}{integer(1) the number of filter queries to undo}
}
\value{
A HCABrowser object with n fewer queries
}
\description{
Undo previous filter queries on a HCABrowser object
}
\examples{
x <- HCABrowser()
x <- x \%>\% filter(organ.text == brain)
x <- x \%>\% filter(organ.text == heart)
x <- x \%>\% filter(organ.text != brain)
#x <- x \%>\% undoEsquery(n = 2)
x
}
