% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utilities.R
\docType{methods}
\name{values,HCABrowser-method}
\alias{values,HCABrowser-method}
\title{List all values for certain fields}
\usage{
\S4method{values}{HCABrowser}(x, fields = c(), ...)
}
\arguments{
\item{x}{An HCABrowser Object.}

\item{fields}{a character vector of fields to display avaiable values for.}

\item{...}{Other arguments.}
}
\description{
List all values for certain fields
}
