% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utilities.R
\docType{methods}
\name{values,ProjectBrowser-method}
\alias{values,ProjectBrowser-method}
\title{List all values for certain fields in a ProjectBrowser Object}
\usage{
\S4method{values}{ProjectBrowser}(x, fields)
}
\arguments{
\item{x}{A ProjectBrowser Object.}

\item{fields}{a character vector of fields to display avaiable values for.}

\item{...}{Other arguments.}
}
\value{
a list of possible values for a filter
}
\description{
List all values for certain fields in a ProjectBrowser Object
}
\examples{
hca <- ProjectBrowser()
vals <- hca \%>\% values('organ')
vals

}
