% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utilities.R
\docType{methods}
\name{fields,HCABrowser-method}
\alias{fields,HCABrowser-method}
\title{List supported fields of an HCABrowser object}
\usage{
\S4method{fields}{HCABrowser}(hca)
}
\arguments{
\item{hca}{An HCABrowser object.}
}
\value{
A tibble indicating fields that can be queried upon.
}
\description{
List supported fields of an HCABrowser object
}
