% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca-api.R
\docType{methods}
\name{hca-api-methods}
\alias{hca-api-methods}
\alias{listBundles,}
\alias{getBundlesCheckout,}
\alias{getBundle,}
\alias{checkoutBundle,}
\alias{getFile,}
\alias{headFile,}
\alias{searchBundles}
\alias{listBundles,HCABrowser-method}
\alias{getBundleCheckout,HCABrowser-method}
\alias{getBundle,HCABrowser-method}
\alias{checkoutBundle,HCABrowser-method}
\alias{getFile,HCABrowser-method}
\alias{headFile,HCABrowser-method}
\alias{searchBundles,HCABrowser-method}
\title{HCA API methods}
\usage{
listBundles(x, ...)
getBunldesCheckout(x, ...)
getBundle(x, ...)
checkoutBundle(x, ...)
getFile(x, ...)
headFile(x, ...)
searchBundles(x, ...)

\S4method{listBundles}{HCABrowser}(x, replica = c("aws", "gcp"), prefix,
  token, per_page = 100, search_after)

\S4method{getBundleCheckout}{HCABrowser}(x, replcia = c("aws", "gcp"),
  checkout_job_id)

\S4method{getBundle}{HCABrowser}(x, uuid, version, replica = c("aws",
  "gcp"), directurls, presignedurls, token, per_page = 500, start_at)

\S4method{checkoutBundle}{HCABrowser}(x, uuid, version,
  replica = c("aws", "gcp"), json_request_body)

\S4method{getFile}{HCABrowser}(x, uuid, replica = c("aws", "gcp"),
  version, token, directurl, content_disposition)

\S4method{headFile}{HCABrowser}(x, uuid, replica = c("aws", "gcp"),
  version)

\S4method{searchBundles}{HCABrowser}(x, json_request_body,
  output_format = c("summary", "raw"), replica = c("aws", "gcp"),
  per_page = 100, search_after)
}
\arguments{
\item{x}{An HCABrowser object that is the subject of the request.}

\item{replica}{character(1). A replica to fetch form. Can either be
set to "aws", "gcp", or "azure".  DEFAULT is "aws".}

\item{token}{\code{Token}. Token to manage retries. End users constructing
queries should not set this parameter. Use \code{get_token()} to generate.}

\item{per_page}{numeric(1). Max number of results to return per page.}

\item{search_after}{character(1). **Search-After-Context**. An internal state
pointer parameter for use with pagination. The API client should not need to
set this parameter directly; it should instead directly fetch the URL given
in the "Link" header.}

\item{checkout_job_id}{character(1). A RFC4122-complliant ID for the checkout
job request.}

\item{uuid}{character(1). A RFC4122-compliant ID for the bundle.}

\item{version}{character(1). Timestamp of bundle creation in RFC3339.}

\item{directurls}{logical(1). Include direct-access URLs in the response.
This is mutually exclusive with the \code{presignedurls} parameter.  DEFAULT
is \code{NULL}.}

\item{presignedurls}{logical(1). Include presigned URLs in the response. This
is mutually exclusive with the directurls parameter.}

\item{output_format}{character(1). Specifies the output format. Either
"summary" or "raw". The default format, "summary", is a list of UUIDs for
bundles that match the query. Set this parameter to "raw" to get the
verbatim JSON metadata for bundles that match the query.}

\item{contents}{list. A list of objects describing links to files, bundles,
other collections, and metadata fragments that are part of the collection.}

\item{creator_uid}{character(1). User ID who is creating this bundle.}

\item{description}{character(1). A long description of the collection,
formatted in Markdown.}

\item{destination}{character(1). User-owned destination storage bucket.}

\item{details}{list. Supplementary JSON metadata for the collection.
(ADD DESCRIPTION OF STRUCTURE)}

\item{es_query}{list. Elasticsearch query. (ADD DESCRIPTION OF STRUCTURE)}

\item{json}{character(1) of a json query to be executed.}

\item{source_url}{character(1). Cloud URL for source data.}
}
\value{
an HCABrowser object
}
\description{
Methods to access the Human Cell Atlas's Data Coordination Platform (HCA DCP)
by means of the platform's REST API.
}
\examples{
hca <- HCABrowser()
#addmore


}
\author{
Daniel Van Twisk
}
