% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HCABrowser-class}
\alias{HCABrowser-class}
\alias{.HCABrowser}
\title{The HCABrowser Class}
\arguments{
\item{url}{character(1) the url of the Human Cell Atlas resource.}

\item{fields_path}{character(1) path to the fields json file.}

\item{per_page}{numeric(1) numbers of pages to view at a time.}
}
\description{
The HCABrowser Class
}
\author{
Daniel Van Twisk
}
