% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{HCABrowser}
\alias{HCABrowser}
\title{The HCABrowser Class}
\usage{
HCABrowser(url = "https://dss.data.humancellatlas.org/v1",
  fields_path = system.file("extdata", "fields_and_values.json", package
  = "HCABrowser"), per_page = 10)
}
\description{
The HCABrowser Class
}
\examples{
hca <- HCABrowser()
hca

}
\author{
Daniel Van Twisk
}
