% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{HCABrowser}
\alias{HCABrowser}
\title{The HCABrowser Class}
\usage{
HCABrowser(host = "dss.data.humancellatlas.org",
  api_url = "https://dss.data.humancellatlas.org/v1/swagger.json",
  per_page = 10)
}
\arguments{
\item{per_page}{numeric(1) numbers of pages to view at a time.}

\item{url}{character(1) the url of the Human Cell Atlas resource.}

\item{fields_path}{character(1) path to the fields json file.}
}
\value{
an HCABrowser object
}
\description{
The HCABrowser Class
}
\examples{
hca <- HCABrowser()
hca
}
\author{
Daniel Van Twisk
}
