% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{HCABrowser}
\alias{HCABrowser}
\title{The HCABrowser Class}
\usage{
HCABrowser(
  host = "dss.data.humancellatlas.org",
  api_url = "https://dss.data.humancellatlas.org/v1/swagger.json",
  per_page = 10,
  authenticate = FALSE
)
}
\arguments{
\item{host}{character(1) path to hca-dcp server}

\item{api_url}{character(1) path to hca-dcp api file}

\item{per_page}{numeric(1) numbers of pages to view at a time.}

\item{authenticate}{logical(1) authenticate? required for actions
that update the HCA data repostiory (rare).}
}
\value{
An HCABrowser object.
}
\description{
The HCABrowser Class
}
\examples{
 hca <- HCABrowser()
 hca

}
\author{
Daniel Van Twisk
}
