% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{ProjectBrowser}
\alias{ProjectBrowser}
\title{The Project Browser Class}
\usage{

  ProjectBrowser(url = "https://service.explore.data.humancellatlas.org/repository/projects",
  per_page = 15)
}
\arguments{
\item{url}{character(1) The url of the Human Cell Atlas.}

\item{per_page}{the number of results to display per request.}
}
\value{
a ProjectBrowser object
}
\description{
A still tentative class that displays Human Cell Atlas
 information by projects.
}
\examples{
pb <- ProjectBrowser()
pb
}
\author{
Daniel Van Twisk
}
