% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{results,HCABrowser-method}
\alias{results,HCABrowser-method}
\title{Obtain search results from a HHACBrowser Object}
\usage{
\S4method{results}{HCABrowser}(object, n = object@per_page,
  all = FALSE)
}
\arguments{
\item{object}{A Human Cell Atlas object}
}
\value{
a tibble
}
\description{
Returns a tibble either showing bundles or files based on whichever is
 activated.
}
