% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{results}
\alias{results}
\alias{results,HCABrowser-method}
\title{Obtain search results from a HCABrowser Object}
\usage{
\S4method{results}{HCABrowser}(object, n = object@per_page,
  all = FALSE, .output_format = c("raw", "summary"))
}
\arguments{
\item{object}{A Human Cell Atlas object.}

\item{n}{numeric(1) number of elements to return.}

\item{all}{logical(1) whether to return all elements.}

\item{.output_format}{unused.}
}
\value{
a tibble
}
\description{
Returns a tibble either showing bundles or files based on whichever is
 activated.
}
