% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{SearchResult}
\alias{SearchResult}
\title{The SearchResults Class}
\usage{
SearchResult(es_query, results, first_hit, last_hit, total_hits)
}
\arguments{
\item{es_query}{A quosure of the current es_query.}

\item{results}{A list of all result from the qeury.}

\item{first_hit}{numeric(1) the first bundle currently shown.}

\item{last_hit}{numeric(1) the last bundle currently shown.}

\item{total_hits}{numeric(1) the number of bundles that can be shown.}
}
\value{
A Search Result object
}
\description{
A glass generated after parsing a search query with
 the method parseToSearchResults. Contains a list of all information
 gleaned from the search query.
}
\examples{
 sr <- new("SearchResult")
 sr

}
