% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{undoEsQuery}
\alias{undoEsQuery}
\alias{undoEsQuery,HCABrowser-method}
\title{Undo previous filter queries on a HCABrowser object}
\usage{
\S4method{undoEsQuery}{HCABrowser}(hca, n = 1L)
}
\arguments{
\item{hca}{A HCABrowser object}

\item{n}{integer(1) the number of filter queries to undo}
}
\value{
A HCABrowser object with n fewer queries
}
\description{
Undo previous filter queries on a HCABrowser object
}
\examples{
hca <- HCABrowser()
hca <- hca \%>\% filter(organ.text == brain)
hca <- hca \%>\% filter(organ.text == heart)
hca <- hca \%>\% filter(organ.text != brain)
#hca <- hca \%>\% undoEsquery(n = 2)
hca
}
